Content Machine/requirements.txt
anthropic>=0.40.0
python-dotenv>=1.0.0
requests>=2.31.0
tavily-python>=0.7.0
streamlit>=1.35.0
